Output the CommsCost report as the underlying object rather than as the final string
